CC-551 fix test

+++ Bundles/Discount/src/Spryker/Zed/Discount/Business/Persistence/DiscountPersist.php
@@ -408,6 +408,10 @@ class DiscountPersist implements DiscountPersistInterface
      */
     protected function deleteDiscountMoneyValues(SpyDiscount $discountEntity): void
     {
+        if (!$discountEntity->getDiscountAmounts()) {
+            return;
+        }
+
         foreach ($discountEntity->getDiscountAmounts() as $discountAmountEntity) {
             $discountAmountEntity->delete();
         }
